+++ branches/jdo2/client/login/src/main/java/com/tll/server/rpc/ForgotPasswordService.java
@@ -51,7 +51,7 @@ public class ForgotPasswordService extends RpcServlet implements IForgotPassword
 			try {
 				final IForgotPasswordHandler handler = context.getForgotPasswordHandler();
 				final IUserRef user = handler.getUserRef(emailAddress);
-				final String rp = handler.resetPassword(user.getId());
+				final String rp = handler.resetPassword(user.getId().longValue());
 				data.put("username", user.getUsername());
 				data.put("emailAddress", user.getEmailAddress());
 				data.put("password", rp);
